+++ src/firewall_rule.rs
@@ -3,24 +3,19 @@ use std::collections::HashSet;
 use std::convert::TryFrom;
 use typed_builder::TypedBuilder;
 use windows::Win32::Foundation::VARIANT_BOOL;
-use windows::Win32::NetworkManagement::WindowsFirewall::{INetFwRule, NetFwRule};
-use windows::Win32::System::Com::CoCreateInstance;
+use windows::Win32::NetworkManagement::WindowsFirewall::INetFwRule;
 use windows::core::BSTR;
 
-use crate::constants::DWCLSCONTEXT;
 use crate::errors::{SetRuleError, WindowsFirewallError};
-use crate::firewall_enums::{Action, Direction, Profile, Protocol};
-use crate::firewall_rule::firewall_address::Address;
-use crate::firewall_rule::firewall_port::Port;
 use crate::utils::{
-    bstr_to_hashset, hashset_to_bstr, hashset_to_variant, into_hashset, is_not_icmp,
-    is_not_tcp_or_udp, variant_to_hashset, with_com_initialized,
+    BstrExt, bstr_to_hashset, hashset_to_bstr, hashset_to_variant, into_hashset,
+    variant_to_hashset, with_rule,
 };
-use crate::windows_firewall::{add_rule_or_update, remove_rule, rule_exists, update_rule};
-use crate::{InterfaceType, add_rule, add_rule_if_not_exists, enable_rule};
 
-pub mod firewall_address;
-pub mod firewall_port;
+pub use self::types::{Action, Address, Direction, InterfaceType, Port, Profile, Protocol};
+
+pub mod operations;
+pub mod types;
 
 /// Represents a rule in the Windows Firewall.
 ///
@@ -138,321 +133,28 @@ pub struct FirewallRule {
     edge_traversal: Option<bool>,
 }
 
-impl FirewallRule {
-    /// Adds a new firewall rule to the system.
-    ///
-    /// This function creates and adds a new firewall rule based on the current instance's properties.
-    /// The rule is registered with the Windows Firewall, applying the specified settings such as
-    /// name, direction, action, and other parameters.
-    ///
-    /// # Arguments
-    ///
-    /// This function does not take any arguments, as it operates on the current instance's fields.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<(), WindowsFirewallError>`](WindowsFirewallError). If the rule is added successfully,
-    /// it returns `Ok(())`. In case of an error (e.g., COM initialization failure or failure to add rule),
-    /// it returns a [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] if there is a failure during:
-    /// - COM initialization [`WindowsFirewallError::CoInitializeExFailed`].
-    /// - Adding the firewall rule.
-    ///
-    /// # Security
-    ///
-    /// ⚠️ This function requires **administrative privileges**.
-    pub fn add(&self) -> Result<(), WindowsFirewallError> {
-        add_rule(self)
-    }
-
-    /// Adds a new firewall rule to the system only if a rule with the same name doesn't exist.
-    ///
-    /// This function first checks if a rule with the given name exists, and if not,
-    /// adds the new rule to the Windows Firewall.
-    ///
-    /// # Arguments
-    ///
-    /// This function does not take any arguments, as it operates on the current instance's fields.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<bool, WindowsFirewallError>`](WindowsFirewallError). If the rule is added successfully,
-    /// it returns `Ok(true)`. If the rule already exists, it returns `Ok(false)`. In case of an error
-    /// (e.g., COM initialization failure or failure to add rule), it returns a [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] if there is a failure during:
-    /// - COM initialization [`WindowsFirewallError::CoInitializeExFailed`].
-    /// - Checking if the rule exists.
-    /// - Adding the firewall rule.
-    ///
-    /// # Security
-    ///
-    /// ⚠️ This function requires **administrative privileges**.
-    pub fn add_if_not_exists(&self) -> Result<bool, WindowsFirewallError> {
-        add_rule_if_not_exists(self)
-    }
-
-    /// Adds a new firewall rule to the system or updates an existing rule with the same name.
-    ///
-    /// This function first checks if a rule with the given name exists. If it does, the function updates
-    /// the existing rule with the new settings. If the rule does not exist, it adds a new rule to the
-    /// Windows Firewall.
-    ///
-    /// # Arguments
-    ///
-    /// This function does not take any arguments, as it operates on the current instance's fields.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<bool, WindowsFirewallError>`](WindowsFirewallError). If the rule is added
-    /// successfully, it returns `Ok(true)`. If the rule already exists and was updated, it returns `Ok(false)`.
-    /// In case of an error (e.g., COM initialization failure, failure to add or update the rule), it returns a
-    /// [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] if there is a failure during:
-    /// - COM initialization [`WindowsFirewallError::CoInitializeExFailed`].
-    /// - Fetching the existing rule or adding the new rule.
-    /// - Updating the firewall rule.
-    ///
-    /// # Security
-    ///
-    /// ⚠️ This function requires **administrative privileges**.
-    pub fn add_or_update(&self) -> Result<bool, WindowsFirewallError> {
-        add_rule_or_update(self)
-    }
-
-    /// Deletes an existing firewall rule from the system.
-    ///
-    /// This function removes a firewall rule identified by its name. Once deleted, the rule
-    /// can no longer be applied, and any associated settings are lost.
-    ///
-    /// # Arguments
-    ///
-    /// This function does not take any arguments, as it operates on the current instance's [`name`](FirewallRule::name) field.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<(), WindowsFirewallError>`](WindowsFirewallError). If the rule is removed successfully,
-    /// it returns `Ok(())`. In case of an error (e.g., COM initialization failure, rule not found),
-    /// it returns a [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] if there is a failure during:
-    /// - COM initialization [`WindowsFirewallError::CoInitializeExFailed`].
-    /// - Removing the rule.
-    ///
-    /// # Security
-    ///
-    /// ⚠️ This function requires **administrative privileges**.
-    pub fn remove(self) -> Result<(), WindowsFirewallError> {
-        remove_rule(&self.name)?;
-        Ok(())
-    }
-
-    /// Updates an existing firewall rule with new settings.
-    ///
-    /// This function modifies an existing firewall rule based on the provided [`settings`](FirewallRuleUpdate).
-    /// It updates various properties such as name, direction, action, protocol, and addresses.
-    /// If the protocol is ICMP (IPv4 or IPv6), the function ensures that local and remote ports
-    /// are cleared, as they are not applicable to ICMP.
-    ///
-    /// # Arguments
-    ///
-    /// * `settings` - A reference to a [`FirewallRuleUpdate`] struct containing the new
-    ///   configuration parameters for the firewall rule.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<(), WindowsFirewallError>`](WindowsFirewallError). If the rule is updated successfully,
-    /// it returns `Ok(())`. In case of an error (e.g., COM initialization failure, rule not found, or failure
-    /// to update the rule), it returns a [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] if there is a failure during:
-    /// - COM initialization [`WindowsFirewallError::CoInitializeExFailed`].
-    /// - Fetching the rule.
-    ///
-    /// # Security
-    ///
-    /// ⚠️ This function requires **administrative privileges**.
-    pub fn update(&mut self, settings: &FirewallRuleUpdate) -> Result<(), WindowsFirewallError> {
-        update_rule(&self.name, settings)?;
-
-        if let Some(name) = &settings.name {
-            self.name = name.clone();
-        }
-        if let Some(direction) = &settings.direction {
-            self.direction = *direction;
-        }
-        if let Some(enabled) = settings.enabled {
-            self.enabled = enabled;
-        }
-        if let Some(action) = &settings.action {
-            self.action = *action;
-        }
-        if let Some(description) = &settings.description {
-            self.description = Some(description.clone());
-        }
-        if let Some(application_name) = &settings.application_name {
-            self.application_name = Some(application_name.clone());
-        }
-        if let Some(service_name) = &settings.service_name {
-            self.service_name = Some(service_name.clone());
-        }
-        if let Some(protocol) = &settings.protocol {
-            if is_not_tcp_or_udp(*protocol) {
-                self.local_ports = None;
-                self.remote_ports = None;
-            }
-            if is_not_icmp(*protocol) {
-                self.icmp_types_and_codes = None;
-            }
-            self.protocol = Some(*protocol);
-        }
-        if let Some(local_ports) = &settings.local_ports {
-            self.local_ports = Some(local_ports.clone());
-        }
-        if let Some(remote_ports) = &settings.remote_ports {
-            self.remote_ports = Some(remote_ports.clone());
-        }
-        if let Some(local_addresses) = &settings.local_addresses {
-            self.local_addresses = Some(local_addresses.clone());
-        }
-        if let Some(remote_addresses) = &settings.remote_addresses {
-            self.remote_addresses = Some(remote_addresses.clone());
-        }
-        if let Some(icmp_types_and_codes) = &settings.icmp_types_and_codes {
-            self.icmp_types_and_codes = Some(icmp_types_and_codes.clone());
-        }
-        if let Some(interfaces) = &settings.interfaces {
-            self.interfaces = Some(interfaces.clone());
-        }
-        if let Some(interface_types) = &settings.interface_types {
-            self.interface_types = Some(interface_types.clone());
-        }
-        if let Some(grouping) = &settings.grouping {
-            self.grouping = Some(grouping.clone());
-        }
-        if let Some(profiles) = &settings.profiles {
-            self.profiles = Some(*profiles);
-        }
-        if let Some(edge_traversal) = &settings.edge_traversal {
-            self.edge_traversal = Some(*edge_traversal);
-        }
-        Ok(())
-    }
-
-    /// Enables or disables an existing firewall rule.
-    ///
-    /// This function modifies the state of a firewall rule based on the `enable` parameter.
-    /// If `enable` is `true`, the rule is enabled; otherwise, it is disabled. The function
-    /// updates the `enabled` field accordingly.
-    ///
-    /// # Arguments
-    ///
-    /// * `enable` - A boolean indicating whether to enable (`true`) or disable (`false`) the rule.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<(), WindowsFirewallError>`](WindowsFirewallError). If the rule is updated successfully,
-    /// it returns `Ok(())`. If an error occurs (e.g., COM initialization failure, rule not found),
-    /// it returns a [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] if there is a failure during:
-    /// - COM initialization [`WindowsFirewallError::CoInitializeExFailed`].
-    /// - Fetching the rule.
-    /// - Enabling or disabling the rule.
-    ///
-    /// # Security
-    ///
-    /// ⚠️ This function requires **administrative privileges**.
-    pub fn enable(&mut self, enable: bool) -> Result<(), WindowsFirewallError> {
-        enable_rule(&self.name, enable)?;
-        self.enabled = enable;
-        Ok(())
-    }
-
-    /// Checks if a firewall rule with the given name exists.
-    ///
-    /// This function initializes COM, creates a firewall policy object, and checks if a rule
-    /// with the specified name exists in the Windows Firewall rules list.
-    ///
-    /// # Arguments
-    ///
-    /// This function does not take any arguments, as it operates on the current instance's [`name`](FirewallRule::name) field.
-    ///
-    /// # Returns
-    ///
-    /// This function returns a [`Result<bool, WindowsFirewallError>`](WindowsFirewallError). If the rule exists, it returns `Ok(true)`,
-    /// otherwise it returns `Ok(false)`. In case of an error (e.g., COM initialization failure or issue
-    /// with firewall policy), it returns a [`WindowsFirewallError`].
-    ///
-    /// # Errors
-    ///
-    /// This function may return a [`WindowsFirewallError`] in case of failures during COM initialization
-    /// or while interacting with the firewall policy object.
-    ///
-    /// # Security
-    ///
-    /// This function does not require administrative privileges.
-    pub fn exists(&self) -> Result<bool, WindowsFirewallError> {
-        rule_exists(&self.name)
-    }
-}
-
 impl TryFrom<INetFwRule> for FirewallRule {
     type Error = WindowsFirewallError;
 
     fn try_from(fw_rule: INetFwRule) -> Result<Self, WindowsFirewallError> {
         unsafe {
             Ok(Self {
-                name: fw_rule.Name().map(|bstr| bstr.to_string())?,
+                name: fw_rule.Name().to_required_string()?,
                 direction: fw_rule.Direction()?.try_into()?,
                 enabled: fw_rule.Enabled()?.into(),
                 action: fw_rule.Action()?.try_into()?,
-                description: fw_rule
-                    .Description()
-                    .ok()
-                    .map(|bstr| bstr.to_string())
-                    .filter(|s| !s.is_empty()),
-                application_name: fw_rule
-                    .ApplicationName()
-                    .ok()
-                    .map(|bstr| bstr.to_string())
-                    .filter(|s| !s.is_empty()),
-                service_name: fw_rule
-                    .ServiceName()
-                    .ok()
-                    .map(|bstr| bstr.to_string())
-                    .filter(|s| !s.is_empty()),
+                description: fw_rule.Description().to_optional_string(),
+                application_name: fw_rule.ApplicationName().to_optional_string(),
+                service_name: fw_rule.ServiceName().to_optional_string(),
                 protocol: fw_rule.Protocol()?.try_into().ok(),
                 local_ports: bstr_to_hashset(fw_rule.LocalPorts()),
                 remote_ports: bstr_to_hashset(fw_rule.RemotePorts()),
                 local_addresses: bstr_to_hashset(fw_rule.LocalAddresses()),
                 remote_addresses: bstr_to_hashset(fw_rule.RemoteAddresses()),
-                icmp_types_and_codes: fw_rule
-                    .IcmpTypesAndCodes()
-                    .ok()
-                    .map(|bstr| bstr.to_string())
-                    .filter(|s| !s.is_empty()),
+                icmp_types_and_codes: fw_rule.IcmpTypesAndCodes().to_optional_string(),
                 interfaces: Some(variant_to_hashset(&fw_rule.Interfaces()?)?),
                 interface_types: bstr_to_hashset(fw_rule.InterfaceTypes()),
-                grouping: fw_rule
-                    .Grouping()
-                    .ok()
-                    .map(|bstr| bstr.to_string())
-                    .filter(|s| !s.is_empty()),
+                grouping: fw_rule.Grouping().to_optional_string(),
                 profiles: fw_rule.Profiles()?.try_into().ok(),
                 edge_traversal: fw_rule.EdgeTraversal().ok().map(VARIANT_BOOL::as_bool),
             })
@@ -464,90 +166,124 @@ impl TryFrom<&FirewallRule> for INetFwRule {
     type Error = WindowsFirewallError;
 
     fn try_from(rule: &FirewallRule) -> Result<Self, WindowsFirewallError> {
-        with_com_initialized(|| unsafe {
-            let fw_rule: Self = CoCreateInstance(&NetFwRule, None, DWCLSCONTEXT)?;
-
-            fw_rule
-                .SetName(&BSTR::from(&rule.name))
-                .map_err(SetRuleError::Name)?;
-            fw_rule
-                .SetDirection(rule.direction.into())
-                .map_err(SetRuleError::Direction)?;
-            fw_rule
-                .SetEnabled(rule.enabled.into())
-                .map_err(SetRuleError::Enabled)?;
-            fw_rule
-                .SetAction(rule.action.into())
-                .map_err(SetRuleError::Action)?;
-            if let Some(ref description) = rule.description {
+        with_rule(|fw_rule| {
+            unsafe {
                 fw_rule
-                    .SetDescription(&BSTR::from(description))
-                    .map_err(SetRuleError::Description)?;
+                    .SetName(&BSTR::from(&rule.name))
+                    .map_err(SetRuleError::Name)
+            }?;
+            unsafe {
+                fw_rule
+                    .SetDirection(rule.direction.into())
+                    .map_err(SetRuleError::Direction)
+            }?;
+            unsafe {
+                fw_rule
+                    .SetEnabled(rule.enabled.into())
+                    .map_err(SetRuleError::Enabled)
+            }?;
+            unsafe {
+                fw_rule
+                    .SetAction(rule.action.into())
+                    .map_err(SetRuleError::Action)
+            }?;
+            if let Some(ref description) = rule.description {
+                unsafe {
+                    fw_rule
+                        .SetDescription(&BSTR::from(description))
+                        .map_err(SetRuleError::Description)
+                }?;
             }
             if let Some(ref app_name) = rule.application_name {
-                fw_rule
-                    .SetApplicationName(&BSTR::from(app_name))
-                    .map_err(SetRuleError::ApplicationName)?;
+                unsafe {
+                    fw_rule
+                        .SetApplicationName(&BSTR::from(app_name))
+                        .map_err(SetRuleError::ApplicationName)
+                }?;
             }
             if let Some(ref service_name) = rule.service_name {
-                fw_rule
-                    .SetServiceName(&BSTR::from(service_name))
-                    .map_err(SetRuleError::ServiceName)?;
+                unsafe {
+                    fw_rule
+                        .SetServiceName(&BSTR::from(service_name))
+                        .map_err(SetRuleError::ServiceName)
+                }?;
             }
             if let Some(protocol) = rule.protocol {
-                fw_rule
-                    .SetProtocol(protocol.into())
-                    .map_err(SetRuleError::Protocol)?;
+                unsafe {
+                    fw_rule
+                        .SetProtocol(protocol.into())
+                        .map_err(SetRuleError::Protocol)
+                }?;
             }
             if let Some(ref local_ports) = rule.local_ports {
-                fw_rule
-                    .SetLocalPorts(&hashset_to_bstr(Some(local_ports)))
-                    .map_err(SetRuleError::LocalPorts)?;
+                unsafe {
+                    fw_rule
+                        .SetLocalPorts(&hashset_to_bstr(Some(local_ports)))
+                        .map_err(SetRuleError::LocalPorts)
+                }?;
             }
             if let Some(ref remote_ports) = rule.remote_ports {
-                fw_rule
-                    .SetRemotePorts(&hashset_to_bstr(Some(remote_ports)))
-                    .map_err(SetRuleError::RemotePorts)?;
+                unsafe {
+                    fw_rule
+                        .SetRemotePorts(&hashset_to_bstr(Some(remote_ports)))
+                        .map_err(SetRuleError::RemotePorts)
+                }?;
             }
             if let Some(ref local_addresses) = rule.local_addresses {
-                fw_rule
-                    .SetLocalAddresses(&hashset_to_bstr(Some(local_addresses)))
-                    .map_err(SetRuleError::LocalAddresses)?;
+                unsafe {
+                    fw_rule
+                        .SetLocalAddresses(&hashset_to_bstr(Some(local_addresses)))
+                        .map_err(SetRuleError::LocalAddresses)
+                }?;
             }
             if let Some(ref remote_addresses) = rule.remote_addresses {
-                fw_rule
-                    .SetRemoteAddresses(&hashset_to_bstr(Some(remote_addresses)))
-                    .map_err(SetRuleError::RemoteAddresses)?;
+                unsafe {
+                    fw_rule
+                        .SetRemoteAddresses(&hashset_to_bstr(Some(remote_addresses)))
+                        .map_err(SetRuleError::RemoteAddresses)
+                }?;
             }
             if let Some(ref icmp_types_and_codes) = rule.icmp_types_and_codes {
-                fw_rule
-                    .SetIcmpTypesAndCodes(&BSTR::from(icmp_types_and_codes))
-                    .map_err(SetRuleError::IcmpTypesAndCodes)?;
+                unsafe {
+                    fw_rule
+                        .SetIcmpTypesAndCodes(&BSTR::from(icmp_types_and_codes))
+                        .map_err(SetRuleError::IcmpTypesAndCodes)
+                }?;
             }
             if let Some(edge_traversal) = rule.edge_traversal {
-                fw_rule
-                    .SetEdgeTraversal(edge_traversal.into())
-                    .map_err(SetRuleError::EdgeTraversal)?;
+                unsafe {
+                    fw_rule
+                        .SetEdgeTraversal(edge_traversal.into())
+                        .map_err(SetRuleError::EdgeTraversal)
+                }?;
             }
             if let Some(ref grouping) = rule.grouping {
-                fw_rule
-                    .SetGrouping(&BSTR::from(grouping))
-                    .map_err(SetRuleError::Grouping)?;
+                unsafe {
+                    fw_rule
+                        .SetGrouping(&BSTR::from(grouping))
+                        .map_err(SetRuleError::Grouping)
+                }?;
             }
             if let Some(ref interface) = rule.interfaces {
-                fw_rule
-                    .SetInterfaces(&hashset_to_variant(interface)?)
-                    .map_err(SetRuleError::Interfaces)?;
+                unsafe {
+                    fw_rule
+                        .SetInterfaces(&hashset_to_variant(interface)?)
+                        .map_err(SetRuleError::Interfaces)
+                }?;
             }
             if let Some(ref interface_types) = rule.interface_types {
-                fw_rule
-                    .SetInterfaceTypes(&hashset_to_bstr(Some(interface_types)))
-                    .map_err(SetRuleError::InterfaceType)?;
+                unsafe {
+                    fw_rule
+                        .SetInterfaceTypes(&hashset_to_bstr(Some(interface_types)))
+                        .map_err(SetRuleError::InterfaceType)
+                }?;
             }
             if let Some(profiles) = rule.profiles {
-                fw_rule
-                    .SetProfiles(profiles.into())
-                    .map_err(SetRuleError::Profiles)?;
+                unsafe {
+                    fw_rule
+                        .SetProfiles(profiles.into())
+                        .map_err(SetRuleError::Profiles)
+                }?;
             }
 
             Ok(fw_rule)
